pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "casestack"
version = "0.1.0"
description = "Turn any document dump into a searchable evidence database"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "rhowardstone"},
]
keywords = ["osint", "document-processing", "ocr", "pipeline", "forensics", "datasette"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Information Analysis",
]
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "rapidfuzz>=3.0",
    "httpx>=0.25",
    "pyyaml>=6.0",
    "datasette>=0.64",
]

[project.optional-dependencies]
ocr = ["docling>=2.0"]
pymupdf = ["pymupdf>=1.24"]
nlp = ["spacy>=3.7"]
ai = ["openai>=1.0"]
ask = ["starlette>=0.27", "uvicorn>=0.24"]
server = ["fastapi>=0.100", "uvicorn[standard]>=0.20", "aiosqlite>=0.20"]
vision = ["pymupdf>=1.24", "Pillow>=10.0"]
transcription = ["faster-whisper>=1.0"]
documents = ["markitdown[all]>=0.1"]
embeddings = ["sentence-transformers>=3.0", "torch>=2.0", "numpy>=1.24"]
captioning = ["transformers>=4.41", "torch>=2.1", "torchvision", "einops", "timm", "Pillow>=10.0", "qwen-vl-utils>=0.0.8"]
all = ["casestack[ocr,pymupdf,nlp,ai,ask,server,vision,transcription,documents,embeddings,captioning]"]
dev = ["pytest>=7.0", "pytest-asyncio>=0.23", "pytest-cov>=4.0", "ruff>=0.4", "mypy>=1.8"]

[project.scripts]
casestack = "casestack.cli:cli"

[project.urls]
Repository = "https://github.com/rhowardstone/casestack"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
casestack = ["static/**/*", "templates/**/*"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
